Organise source files into folders
